Use short weekday
Means all columns are the same width (3 chars), which makes the diagonal strikethrough look less weird (when there were different size cells, the angles would be different, which looked off)

+++ src/App.tsx
@@ -94,7 +94,7 @@ function getCalendar(currentDate: Date): JSX.Element {
 
     const dotw = thisDate.getDay();
 
-    dayNames[dotw] = thisDate.toLocaleString('default', { weekday: 'long' });
+    dayNames[dotw] = thisDate.toLocaleString('default', { weekday: 'short' });
 
     if (dotw === 0) {
       // start new row
